mainwindow: Add tape loading and saving operations
The file menu is now replaced with a 'Tape' menu. The tapes can be
loaded and saved from here. There is also a tape rewind option. The
dirty state of the tape buffer is checked before creating a new tape
or quitting the emulator. If the buffer is dirty then the user is
prompted for confirmation first.