Add a 'full speed' option
This adds a menu item to set 'full speed' mode. In full speed mode the
frame source is replaced with an idle timeout so that the Electron
will run as fast as possible rather than trying to run at an accurate
speed. The new frame signal is still only emitted at 50fps however so
not every frame generated by the electron will be drawn.