electron: Count key time in frames instead of scans
Instead of holding down the key for a certain number of scans of the
keyboard, it now keeps it held down for a certain number of frames since
the key was added to the queue. This fixes a problem where the escape
key couldn’t be queued because the OS ROM scans that differently,
probably because it is in the same line as the modifiers.