streamer: Use a ring buffer for frames
ffmpeg doesn’t become ready for writing at regular intervals. Instead,
it reads a bunch of frames and then blocks while processing them before
being ready to read a bunch more. This was having the effect that the
emulator was not recording any frames until ffmpeg was ready and then it
would just send a bunch of identical frames. To fix this, the emulator
now keeps its own ring buffer of frames and adds to that even when
ffmpeg is not ready to receive data. That way when it is ready to
receive a bunch of frames they will all have been taken at different
times.